WebGIS/WebGIS Project/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==5.0.6
django-geojson==4.1.0
django-leaflet==0.30.0
GDAL-3.8.4-cp312-cp312-win_amd64.whl
geopandas==1.0.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
